# docs/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Clock Bucks API - Meeting Cost Calculator",
    "description": "# Clock Bucks API - Meeting Cost Calculator\n\nA comprehensive backend service that calculates the cost of meetings based on participant salaries and hourly rates.\n\n## Features\n- **Meeting Management**: Create, update, and track meetings with cost calculations\n- **Participant Management**: Manage participants with roles and hourly rates\n- **Cost Calculations**: Real-time cost tracking during meetings\n- **Statistics**: Meeting analytics and cost reporting\n- **Data Persistence**: File-based storage with plans for database integration\n\n## Use Cases\n- Track meeting costs for budget planning\n- Optimize meeting efficiency by understanding cost implications\n- Generate cost reports for management and accounting\n- Real-time cost awareness during ongoing meetings\n\n## Architecture\n- **FastAPI**: High-performance async Python web framework\n- **Pydantic**: Data validation and serialization\n- **Repository Pattern**: Clean separation of data access logic\n- **File Storage**: JSON-based persistence (interim solution)\n- **Comprehensive Testing**: pytest-based test suite\n\n## Contact\n- **Developer**: Henrique Realinho\n- **Repository**: https://github.com/hrealinho/clockbucks",
    "version": "1.0.0",
    "contact": {
      "name": "Clock Bucks API Support",
      "url": "https://github.com/hrealinho/clockbucks",
      "email": "support@clockbucks.com"
    },
...
  "servers": [
...
  "paths": {
...
    "/api/v1/participants": {
      "post": {
        "tags": ["participants"],
        "summary": "Create New Participant",
        "description": "Create a new participant with their role and hourly rate information for meeting cost calculations.",
        "operationId": "create_participant",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParticipantCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Participant created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid participant data"
          },
          "422": {
            "description": "Validation error"
          }
        }
      },
      "get": {
        "tags": ["participants"],
        "summary": "List All Participants",
        "description": "Retrieve all participants with their current hourly rates and role information.",
        "operationId": "get_all_participants",
        "responses": {
          "200": {
            "description": "Participants retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ParticipantResponse"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Participant": {
        "type": "object",
        "required": ["name", "hourly_rate"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique participant identifier"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "description": "Participant's full name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Participant's email address"
          },
          "hourly_rate": {
            "type": "number",
            "minimum": 0,
            "description": "Hourly rate in USD"
          },
          "role": {
            "type": "string",
            "description": "Job title or role"
          },
          "department": {
            "type": "string",
            "description": "Department or team"
          }
        },
        "example": {
          "name": "John Doe",
          "email": "john.doe@company.com",
          "hourly_rate": 75.0,
          "role": "Senior Developer",
          "department": "Engineering"
        }
      },
      "ParticipantCreate": {
        "type": "object",
        "required": ["name", "hourly_rate"],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "description": "Participant's full name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Participant's email address"
          },
          "hourly_rate": {
            "type": "number",
            "minimum": 0,
            "description": "Hourly rate in USD"
          },
          "role": {
            "type": "string",
            "description": "Job title or role"
          },
          "department": {
            "type": "string",
            "description": "Department or team"
          }
        }
      },
      "ParticipantResponse": {
        "allOf": [
          {"$ref": "#/components/schemas/Participant"},
          {
            "type": "object",
            "required": ["id", "created_at"],
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "description": "Unique participant identifier"
              },
              "created_at": {
                "type": "string",
                "format": "date-time",
                "description": "Account creation timestamp"
              }
            }
          }
        ]
      },
...
    }
  }
}